Small refactor in PlaceOfferSection
