+++ src/components/PlaceOfferSection/PlaceOfferContent/PlaceOfferContent.tsx
@@ -1,17 +1,15 @@
 import { FC, useMemo } from 'react'
 
 import { useWallet } from '@solana/wallet-adapter-react'
-import { useQuery } from '@tanstack/react-query'
 import { chain, sortBy } from 'lodash'
 
 import { InputCounter, InputErrorMessage, NumericInputField } from '@banx/components/inputs'
 
-import { fetchLenderLoansByCertainOffer } from '@banx/api/core'
 import { convertOffersToSimple } from '@banx/pages/BorrowPage/helpers'
 
 import { BorrowerMessage } from '../components'
 import { getUpdatedBondOffer } from '../helpers'
-import { OfferParams } from '../hooks'
+import { OfferParams, useLenderLoans } from '../hooks'
 import { ActionsButtons, Summary } from './components'
 import {
   convertLoanToMark,
@@ -41,7 +39,6 @@ const PlaceOfferContent: FC<OfferParams> = ({
   syntheticOffer,
 }) => {
   const { connected } = useWallet()
-  const disabled = !connected
 
   const showBorrowerMessage = !offerErrorMessage && !!offerSize
   const disablePlaceOffer = !!offerErrorMessage || !offerSize
@@ -96,14 +93,14 @@ const PlaceOfferContent: FC<OfferParams> = ({
           value={loanValue}
           onChange={onLoanValueChange}
           className={styles.numericField}
-          disabled={disabled}
+          disabled={!connected}
         />
         {onDeltaValueChange && (
           <NumericInputField
             label="Avg Delta"
             onChange={onDeltaValueChange}
             value={deltaValue}
-            disabled={disabled}
+            disabled={!connected}
             tooltipText="The average difference between loans taken from this pool given 100% utilization. For example: initialOffer: 1 SOL, delta 0.2 SOL, number of offers 2. The loans can be either the max 1, 0.8; or 0.2, 0.4, 0.4, 0,6, 0.1, 0.1. In both cases the average delta is 0.2. And the sum of loans is same"
           />
         )}
@@ -111,7 +108,7 @@ const PlaceOfferContent: FC<OfferParams> = ({
           label="Number of offers"
           onChange={onLoanAmountChange}
           value={loansAmount}
-          disabled={disabled}
+          disabled={!connected}
         />
       </div>
       <div className={styles.messageContainer}>
@@ -139,31 +136,3 @@ const PlaceOfferContent: FC<OfferParams> = ({
 }
 
 export default PlaceOfferContent
-
-export const useLenderLoans = (offerPubKey: string) => {
-  const { publicKey } = useWallet()
-
-  const { data, isLoading, refetch } = useQuery(
-    ['lenderLoans', publicKey?.toBase58(), offerPubKey],
-    () =>
-      fetchLenderLoansByCertainOffer({ walletPublicKey: publicKey?.toBase58() || '', offerPubKey }),
-    {
-      staleTime: 60_000,
-      refetchInterval: 5_000,
-      refetchOnWindowFocus: false,
-    },
-  )
-
-  const lenderLoans = useMemo(() => {
-    if (!data) return []
-
-    return data.flatMap(({ loans }) => loans)
-  }, [data])
-
-  return {
-    data: data ?? [],
-    lenderLoans,
-    isLoading,
-    refetch,
-  }
-}
