+++ src/components/PlaceOfferSection/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useOfferTransactions'
 export * from './usePlaceOffer'
+export * from './useLenderLoans'
